requirements.txt
argh==0.23.2
azure==1.0.3
azure-common==1.0.0
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
coverage==3.6
Django==1.9.4
django-nose==1.2
django-redis-cache==1.6.5
djangorestframework==3.3.2
docutils==0.12
futures==3.0.5
Markdown==2.1.1
mock==1.0.1
nose==1.3.0
ordereddict==1.1
pydocumentdb==1.5.0
python-dateutil==2.4.2
python-mimeparse==1.5.1
PyYAML==3.11
redis==2.10.5
requests==2.9.1
six==1.10.0
wheel==0.29.0
git+https://github.com/mrcabellom/django-rest-swagger.git@master#egg=django-rest-swagger
